WAL no. 3: recycle log files
Recycle log files when switching over to the next file.

When switching, if the truncation index (pointing to the latest
file with a truncation point) is ahead of the first log file,
recycle the first log file into the next one.

If the first log file has the truncation index, create a new
file and let it grow.

In the normal case, when truncation is frequent, there will be
only two log files in the stead state.

Signed-off-by: Yoav Tock <tock@il.ibm.com>