Test with request timeouts
Improve the simulated network to allow forwarding a request
to the leader.

Expose batch-size, batch-timeout, and number of nodes as a
network option.

A test where clients submit each of their requests to a single
node, spread  across all nodes. The leader (node 0) will propose
its requests first. Next the requests in the followers will
timeout. This will cause them to forward the requests to the
leader. The leader will then propose these requests.

This test uses a shorter timeout than default, and batch-size>1.

Signed-off-by: Yoav Tock <tock@il.ibm.com>