Create pool with timeout handler
Currently, the request pool has a SetTimeoutHandler method
which is never called in batcher tests, which leads
to null pointer panic when running with -count 5.

This is a by-product of the fact that one can create a pool without
a timeout handler.

I removed the said method and expanded the constructor so that
you have to pass in a timeout handler.

Signed-off-by: yacovm <yacovm@il.ibm.com>