Check source node is disconnected when sending
Currently when we send to a remote node that is disconnected,
we only check whether that node is disconnected.

However if a node is a leader and it is disconnected, it is still
able to send heartbeats to everyone because none of the followers
are disconnected in case only the leader is disconnected.

In this commit we also add a check if the source node is disconnected.

Signed-off-by: yacovm <yacovm@il.ibm.com>