Abort Delivery if controller stopped
When stopping the controller it might be in the middle of deliverin
to the application, and the stop operation would wait for all
goroutines to wait, which will never happen because there might be
no one reading from the delivery channel at the time.

This change simply makes the controller abort delivery to the app
if it is stopped.

Signed-off-by: yacovm <yacovm@il.ibm.com>