Ignore heartbeats from leaders far ahead
This change makes the heartbeat monitor disregard heartbeats
from leaders that are far away as long as its view is functioning.

This is in order to eventually trigger a sync of a node that
is too far behind but no new proposals arrive and thus it lacks
the information to figure out it is far behind.

As part of the change set, the leader now also sends its proposal sequence
in the heart-beat.

Signed-off-by: yacovm <yacovm@il.ibm.com>