Re-order leader forward logic
This commit makes transactions forwarding to leader first schedule
the leader timeout routine and then send it to the leader, and not the other way around.

This is to prevent a case where we get blocked when sending to the leader,
and never schedule the 2nd timeout routine and thus never remove the transaction from the request pool,
which makes the semaphore count of the request pool reach zero and block all requests.

We also change accordingly, the 2nd timeout - first schedule the removal
and then complain about the leader.

Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>