Fail fast if request already exists in pool
When adding a new request to the pool, there is no need to wait on the semaphore
in case the request already exists, as it is already scheduled to be either picked
by the batcher or to be forwarded to the leader.

This commit returns early in case the request that is about to be added, already exists
in the request pool.

Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>