Restart request pool timers when leader rotates
When the leader rotates due to a block commit, any request that has fired a first timeout
is waiting either for the request to be included in some block, or it will trigger
lobbying for a view change.

This is wrong, because the timeout might have originated from a previous leader and this leader
might be honest but just didn't have enough time to include the request in a block.

This commit resets the timeout upon every leader rotation, to give a fair amount of time
for all leader nodes to schedule pending requests.

Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>