Delete in-flght when syncing past the in-flight sequence
In case a node syncs past its last recorded in-flight proposal, it needs to delete the in-flight because the rest of the nodes
have progressed further than this in-flight proposal.

Hence, it will never need to initialize a view to consent on the in-flight proposal since during the sync it has already committed this proposal.

Even if there are some nodes that still consider the response to be in-flight, the current node can make them commit this proposal as it has corresponding
signatures from the decision of this proposal.

Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>