Properly handle ticks when waiting for In-Flight view to finish
Due to a bug, When the view changer initializes an in-flight view it only waits for a single tick
before resuming view change standard operation.

Another part of the bug is that it anyway returns success, which in turn makes the sequence of the controller
be initialized to an old sequence.

Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>