Recording infrastructure
This commit adds infrastructure for recording the operation of the consensus library
and then replaying it.

Currently only Sync() is implemented, but eventually more capabilities
will be added such as recording messages received, and initial state.

A recording does not record sensitive data, but instead it sanitizes it.
The recording output can then be sent to developers to investigate
and reproduce bugs, in order to solve them.

Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>