Properly handle committing a decision during a view change while having in-flight decision
When committing a decision during a view change, if the node has an
in-flight proposal, it will commit the decision as part of processing a view data
message from a more up-to-date node.

However, the current code crashes once it gets a second view data message, because
its in-flight proposal is now behind the latest decision, which is an illegal state.

This patch makes the code be aware of the possibility of committing a decision while
being in a view change, and then allowing the in-flight proposal be one sequence
behind because of it.

Signed-off-by: Artem Barger <artem@bargr.net>