Use state collector when sync is behind
When sync is behind (returning a previous view number) we return early and do not call fetchState.
However, there might be multiple view changes the sync is not aware of, and therefore it is behind.
So in this case we should call fetchState anyway to get updates about possible view changes.

Signed-off-by: Hagar Meir <hagar.meir@ibm.com>